+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add missing scico palettes. (#168)
 * Added new palettes from ButterflyColors, feathers, ggsci, khroma, PrettyCols, and tayloRswift. (#182)
 * 4 palettes have been added from blueycolors package. (#155)
+* 5 palettes have been added from amerika package. (#158)
 
 # paletteer 1.6.0